fix: pair_lock_termination returns bool, not float
RuntimeError: Float can't be cast to Bool — termination manager
expects bool tensor. Changed terminate.float() to terminate.

+++ source/spot_micro_rl/spot_micro_rl/tasks/manager_based/spot_micro_rl/mdp/rewards.py
@@ -2528,7 +2528,7 @@ def prolonged_pair_lock_termination(
     # grace 이후 + consecutive 초과 시 termination
     in_grace = env.episode_length_buf < grace_steps
     terminate = (~in_grace) & (env._pair_lock_count >= consecutive_steps)
-    return terminate.float()
+    return terminate
 
 
 def diagonal_pair_separation_reward(
